Issue #52: Integrate video calibration from scikit-surgerycalibration 0.0.7
