--- bardGrabVideoImages.py
@@ -1,8 +1,0 @@
-#!/usr/bin/python
-#  -*- coding: utf-8 -*-
-import sys
-
-from sksurgerybard.ui.bard_grab_video_images_command_line import main
-
-if __name__ == "__main__":
-    sys.exit(main(sys.argv[1:]))
