+++ bardVideoCalibration.py
@@ -2,7 +2,7 @@
 #  -*- coding: utf-8 -*-
 import sys
 
-from sksurgerybard.ui.bard_camera_calibration_command_line import main
+from sksurgerycalibration.ui.video_calibration_command_line import main
 
 if __name__ == "__main__":
     sys.exit(main(sys.argv[1:]))
