+++ config/video_calib_chessboard.json
@@ -1,0 +1,7 @@
+{
+    "method": "chessboard",
+    "source": 0,
+    "corners": [14, 10],
+    "square size in mm": 3,
+    "minimum number of views": 5
+}
