+++ doc/.gitignore
@@ -3,3 +3,4 @@
 /modules.rst
 /versioneer.rst
 /build
+bard*.rst
